feat: create every second new qr code and send it via websocket
